+++ clang-tools-extra/clang-tidy/readability/ReadabilityTidyModule.cpp
@@ -46,7 +46,7 @@
 #include "RedundantDeclarationCheck.h"
 #include "RedundantFunctionPtrDereferenceCheck.h"
 #include "RedundantInlineSpecifierCheck.h"
-#include "RedundantLambdaParenthesesCheck.h"
+#include "RedundantLambdaParameterListCheck.h"
 #include "RedundantMemberInitCheck.h"
 #include "RedundantParenthesesCheck.h"
 #include "RedundantPreprocessorCheck.h"
@@ -144,8 +144,8 @@ public:
         "readability-redundant-casting");
     CheckFactories.registerCheck<RedundantFunctionPtrDereferenceCheck>(
         "readability-redundant-function-ptr-dereference");
-    CheckFactories.registerCheck<RedundantLambdaParenthesesCheck>(
-        "readability-redundant-lambda-parentheses");
+    CheckFactories.registerCheck<RedundantLambdaParameterListCheck>(
+        "readability-redundant-lambda-parameter-list");
     CheckFactories.registerCheck<RedundantMemberInitCheck>(
         "readability-redundant-member-init");
     CheckFactories.registerCheck<RedundantParenthesesCheck>(