+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -163,8 +163,8 @@ New checks
   Suggests insertion of ``std::move(...)`` to turn copy assignment operator
   calls into move assignment ones, when deemed valid and profitable.
 
-- New :doc:`readability-redundant-lambda-parentheses
-  <clang-tidy/checks/readability/redundant-lambda-parentheses>` check.
+- New :doc:`readability-redundant-lambda-parameter-list
+  <clang-tidy/checks/readability/redundant-lambda-parameter-list>` check.
 
   Finds lambda expressions with a redundant empty parameter list and removes it.